server/code/requirements.txt
aniso8601==8.1.0

dnspython==2.0.0
email-validator==1.1.2
flask-mongoengine==1.0.0
Flask-WTF==0.14.3
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Faker==5.0.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.21.0
pymongo==3.11.2
WTForms==2.3.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
six==1.15.0
text-unidecode==1.3
urllib3==1.26.2
Werkzeug==1.0.1
